Adding OAS support for Film entity (#34)
* Adding OAS support for Film entity
* Fixing issue applying multiple OAS in the same ms
* Adding changes in jpa module
* Adding changes in jooq module
* Adding changes in jdbc module
* Adding changes in data-jdbc module

+++ spring-data-jdbc/src/main/java/info/jab/ms/repository/FilmQueryProjection.java
@@ -1,2 +1,2 @@
 package info.jab.ms.repository;
-public record FilmQueryProjection(Integer FILM_ID, String TITLE) {}
+public record FilmQueryProjection(Long FILM_ID, String TITLE) {}
